Refactors DB call from component to Page
